requirements.txt
streamlit
streamlit-authenticator
pyyaml
pandas
plotly
bcrypt
streamlit-option-menu